Adding dryRun parameter to builds

+++ publish/src/commands/deploy.js
@@ -64,6 +64,7 @@ const deploy = async ({
 	oracleDepot,
 	privateKey,
 	yes,
+	dryRun = false,
 } = {}) => {
 	ensureNetwork(network);
 	ensureDeploymentPath(deploymentPath);
@@ -274,6 +275,7 @@ const deploy = async ({
 	}
 
 	parameterNotice({
+		'Dry Run': dryRun,
 		Network: network,
 		'Gas price to use': `${gasPrice} GWEI`,
 		'Deployment Path': new RegExp(network, 'gi').test(deploymentPath)
@@ -325,7 +327,7 @@ const deploy = async ({
 	const newContractsDeployed = [];
 	// force flag indicates to deploy even when no config for the entry (useful for new synths)
 	const deployContract = async ({ name, source = name, args, deps, force = false }) => {
-		const deployedContract = await deployer.deploy({ name, source, args, deps, force });
+		const deployedContract = await deployer.deploy({ name, source, args, deps, force, dryRun });
 		if (!deployedContract) {
 			return;
 		}
@@ -354,11 +356,13 @@ const deploy = async ({
 			bytecode: compiled[source].evm.bytecode.object,
 			abi: compiled[source].abi,
 		};
-		fs.writeFileSync(deploymentFile, stringify(deployment));
+		if (!dryRun) {
+			fs.writeFileSync(deploymentFile, stringify(deployment));
+		}
 
 		// now update the flags to indicate it no longer needs deployment,
 		// ignoring this step for local, which wants a full deployment by default
-		if (network !== 'local') {
+		if (network !== 'local' && !dryRun) {
 			updatedConfig[name] = { deploy: false };
 			fs.writeFileSync(configFile, stringify(updatedConfig));
 		}
@@ -390,6 +394,7 @@ const deploy = async ({
 			etherscanLinkPrefix,
 			ownerActions,
 			ownerActionsFile,
+			dryRun,
 		});
 
 	await deployContract({
@@ -767,7 +772,7 @@ const deploy = async ({
 	}
 
 	// Read Synthetix Proxy address
-	const synthetixProxyAddress = await synthetix.methods.proxy().call();
+	const synthetixProxyAddress = proxySynthetix ? proxySynthetix.options.address : '';
 
 	if (supplySchedule && synthetix) {
 		await runStep({
@@ -1327,6 +1332,10 @@ module.exports = {
 			.option(
 				'-p, --oracle-depot <value>',
 				'The address of the depot oracle for this network (default is use existing)'
+			)
+			.option(
+				'-r, --dry-run',
+				'If enabled, will not run any transactions but merely report on them.'
 			)
 			.option(
 				'-v, --private-key [value]',
